requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
numpy>=1.26.0
tqdm>=4.65.0
scikit-learn>=1.3.2
openai>=1.0.0
rapidfuzz>=2.0.0
faiss-cpu==1.13.0
redis>=4.5.5